make output as string not list

+++ test_classification/custom_trainer.py
@@ -8,11 +8,11 @@ class CustomTrainer(Trainer):
 
         # Forward Pass
         outputs = model(**inputs)
-        logits = outputs.get("logits")
+        logits = outputs.get("logits").float()
         logits = logits.float()
-
+        labels = labels.long()
         # Compute Custom Loss
-        loss_fct = nn.CrossEntropyLoss(weight=torch.tensor(self.class_weights).to(device=self.device))
+        loss_fct = nn.CrossEntropyLoss(weight=torch.tensor(self.class_weights, dtype=torch.float32).to(device=self.device))
         loss = loss_fct(logits.view(-1, self.model.config.num_labels), labels.view(-1))
         return (loss, outputs) if return_outputs else loss
